Move workspace handling to separate file and add extend support
